@@ -8,23 +8,29 @@
     "build": "tsc",
     "watch": "tsc -w",
     "cdk": "cdk --profile tp-access",
-    "cdk:diff": "cdk diff --profile tp-access"
+    "cdk:diff": "cdk diff --profile tp-access",
+    "deploy:all": "yarn build && cdk --profile tp-access deloy Tp*",
+    "deploy:gql": "yarn build && cdk --profile tp-access deploy TpGqlStaging"
   },
   "devDependencies": {
-    "@aws-cdk/aws-iam": "^0.32.0",
-    "@aws-cdk/aws-s3": "^0.32.0",
-    "@types/node": "8.10.45",
-    "aws-cdk": "^0.32.0",
+    "@types/node": "12.12.8",
+    "aws-cdk": "^1.16.3",
     "ts-node": "^8.1.0",
-    "typescript": "^3.3.3333"
+    "typescript": "^3.7.2"
   },
   "dependencies": {
-    "@aws-cdk/aws-certificatemanager": "^0.32.0",
-    "@aws-cdk/aws-cloudfront": "^0.32.0",
-    "@aws-cdk/aws-route53": "^0.32.0",
-    "@aws-cdk/aws-route53-targets": "^0.32.0",
-    "@aws-cdk/aws-ssm": "^0.32.0",
-    "@aws-cdk/cdk": "^0.32.0",
+    "@aws-cdk/aws-certificatemanager": "^1.16.3",
+    "@aws-cdk/aws-cloudfront": "^1.16.3",
+    "@aws-cdk/aws-ec2": "^1.16.3",
+    "@aws-cdk/aws-ecr": "^1.16.3",
+    "@aws-cdk/aws-ecs": "^1.16.3",
+    "@aws-cdk/aws-ecs-patterns": "^1.16.3",
+    "@aws-cdk/aws-iam": "^1.16.3",
+    "@aws-cdk/aws-route53": "^1.16.3",
+    "@aws-cdk/aws-route53-targets": "^1.16.3",
+    "@aws-cdk/aws-s3": "^1.16.3",
+    "@aws-cdk/aws-ssm": "^1.16.3",
+    "@aws-cdk/core": "^1.16.3",
     "source-map-support": "^0.5.9"
   }
 }
