@@ -1,9 +1,11 @@
-import cdk = require('@aws-cdk/cdk');
+import cdk = require('@aws-cdk/core');
 import s3 = require('@aws-cdk/aws-s3');
 import iam = require('@aws-cdk/aws-iam');
 import cloudfront = require('@aws-cdk/aws-cloudfront');
 import route53 = require('@aws-cdk/aws-route53');
 import targets = require('@aws-cdk/aws-route53-targets');
+
+import { Construct } from '@aws-cdk/core';
 
 import ssm = require('@aws-cdk/aws-ssm');
 
@@ -13,25 +15,28 @@ const subDomain = 'images';
 
 const siteDomain = `${subDomain}.${domainName}`;
 
-export class TypePartyStack extends cdk.Stack {
-  constructor(scope: cdk.Construct, id: string, props?: cdk.StackProps) {
-    super(scope, id, props);
+export class TpImages extends Construct {
+  constructor(parent: Construct, name: string) {
+    super(parent, name);
 
     const tpImageBucket = new s3.Bucket(this, 'TPImageBucket', {
       bucketName: siteDomain,
-      removalPolicy: cdk.RemovalPolicy.Destroy,
+      removalPolicy: cdk.RemovalPolicy.DESTROY,
     });
 
     const tpImageBucketPolicy = new s3.BucketPolicy(this, 'TPImageBucketPolicy', {
       bucket: tpImageBucket,
     });
 
-    tpImageBucketPolicy.document.addStatement(
-      new iam.PolicyStatement(iam.PolicyStatementEffect.Allow)
-        .addAction('s3:*')
-        .addResource(tpImageBucket.bucketArn)
-        .addAccountRootPrincipal()
+    tpImageBucketPolicy.document.addStatements(
+      new iam.PolicyStatement({
+        effect: iam.Effect.ALLOW,
+        resources: [tpImageBucket.bucketArn],
+        actions: ['s3:*'],
+        principals: [new iam.AccountRootPrincipal()],
+      })
     );
+
     const cfOriginAccessIdentity = new cloudfront.CfnCloudFrontOriginAccessIdentity(
       this,
       'tpImageOAI',
@@ -42,13 +47,14 @@ export class TypePartyStack extends cdk.Stack {
       }
     );
 
-    const arnCertificate = new ssm.ParameterStoreString(this, 'ArnCertificate', {
-      parameterName: `CertifcateArn-${domainName}`,
-    });
+    const arnCertificate = ssm.StringParameter.valueForStringParameter(
+      this,
+      `CertifcateArn-${domainName}`
+    );
 
-    const zone = new route53.HostedZoneProvider(this, {
+    const zone = route53.HostedZone.fromLookup(this, 'Zone', {
       domainName,
-    }).findAndImport(this, 'Zone');
+    });
 
     const distribution = new cloudfront.CloudFrontWebDistribution(
       this,
@@ -56,9 +62,9 @@ export class TypePartyStack extends cdk.Stack {
       {
         aliasConfiguration: {
           names: [`images.${domainName}`],
-          acmCertRef: arnCertificate.stringValue,
+          acmCertRef: arnCertificate,
           sslMethod: cloudfront.SSLMethod.SNI,
-          securityPolicy: cloudfront.SecurityPolicyProtocol.TLSv1_1_2016,
+          securityPolicy: cloudfront.SecurityPolicyProtocol.TLS_V1_1_2016,
         },
         originConfigs: [
           {
@@ -76,9 +82,11 @@ export class TypePartyStack extends cdk.Stack {
       }
     );
 
-    new route53.AliasRecord(this, 'ImagesARecord', {
+    new route53.ARecord(this, 'ImagesARecord', {
       recordName: siteDomain,
-      target: new targets.CloudFrontTarget(distribution),
+      target: route53.AddressRecordTarget.fromAlias(
+        new targets.CloudFrontTarget(distribution)
+      ),
       zone,
     });
     // The code that defines your stack goes here
